Migration of tests in finagle-thrift and finagle-stream to ScalaTest
Problem

These two packages contain tests written in Specs2 and ScalaTest.

Solution

All tests in these packages were migrated to ScalaTest.  Additionally, the ignore facility
in ScalaTest was used to message when flaky tests are disable.

Result

These packages now conform to the Finagle testing stadard.  Additionally, the ignore facility
in ScalaTest was used to message when flaky tests are disabled.
