+++ finagle-thrift/src/test/scala/com/twitter/finagle/thrift/EndToEndTest.scala
@@ -73,7 +73,7 @@ class EndToEndTest extends FunSuite with ThriftTest with Eventually with BeforeA
     }
   }
 
-  testThrift("propagate Dtab") { (client, tracer) =>
+  skipTestThrift("propagate Dtab") { (client, tracer) =>
     Dtab.unwind {
       Dtab.read("/a=>/b; /b=>/$/inet/google.com/80")
       val clientDtab = Await.result(client.show_me_your_dtab())
@@ -117,7 +117,7 @@ class EndToEndTest extends FunSuite with ThriftTest with Eventually with BeforeA
     assert(bytes.toSeq != decoded.toSeq, "Add JSON support back")
   }
 
-  testThrift("end-to-end tracing potpourri") { (client, tracer) =>
+  skipTestThrift("end-to-end tracing potpourri") { (client, tracer) =>
     Trace.unwind {
       Trace.setId(Trace.nextId)  // set an ID so we don't use the default one
       assert(Await.result(client.multiply(10, 30)) === 300)
